updated to new gdnim and fixed paddle reload

+++ tasks.nim
@@ -1,18 +1,29 @@
 from sequtils import toSeq, filter, mapIt
-import times, anycase, threadpool,strutils, osproc
+import anycase, threadpool
 
-const nim_template = """
-import godot
-import godotapi / [$3]
-import hot
+const script_nim_template = """
+import gdnim, godotapi / [$3]
 
 gdobj $2 of $4:
 
-  method enter_tree() =
-    discard register($1)
-
   proc hot_unload():seq[byte] {.gdExport.} =
     self.queue_free()
+    #save()
+
+  method enter_tree() =
+    discard register($1)#?.load()
+"""
+
+const tool_nim_template = """
+import godot, godotapi / [editor_plugin, resource_loader]
+
+gdobj($2 of EditorPlugin, tool):
+
+  method enter_tree() =
+    discard
+
+  method exit_tree() =
+    discard
 """
 
 const gdns_template = """
@@ -41,10 +52,22 @@ const tscn_template = """
 script = ExtResource( 1 )
 """
 
+const plugin_cfg_template = """
+[plugin]
+
+name="$1"
+description=""
+author=""
+version=""
+script="$1.gdns"
+
+"""
+
 let appDir = config.getSectionValue("Dir", "app")
 let compsDir = config.getSectionValue("Dir", "comps")
 let depsDir = config.getSectionValue("Dir", "deps")
 let depsGodotDir = config.getSectionValue("Dir", "deps_godot")
+let gdpathFlags = &"--path:gdnim --path:{depsDir} --path:{depsDir}/{depsGodotDir} "
 
 # generated files
 let baseDllDir = config.getSectionValue("App", "dll")
@@ -72,8 +95,12 @@ let dllExt = case hostOS
                else: "unknown"
 
 proc genGdns(name:string) =
-  let gdns = &"{gdnsDir}/{name}.gdns"
-  let comp = &"{compsDir}/{name}.nim"
+  var comp = &"{compsDir}/{name}.nim"
+  var gdns = &"{gdnsDir}/{name}.gdns"
+  if not fileExists(comp):
+    comp = &"{compsDir}/tools/{name}.nim"
+    gdns = &"{appDir}/addons/{name}/{name}.gdns"
+
   if fileExists(comp) and not fileExists(gdns):
     var f = open(gdns, fmWrite)
     f.write(gdns_template % [name, name.pascal, relativePath(dllDir, appDir)])
@@ -82,25 +109,6 @@ proc genGdns(name:string) =
 
 proc execOrQuit(command:string) =
   if execShellCmd(command) != 0: quit(QuitFailure)
-
-proc checkPrereq(packageName, sourceName:string, verbose:bool = true) =
-  var (output, exitCode) = execCmdEx(&"nimble path {packageName}")
-  if exitCode != 0:
-    echo &"{packageName} is not installed. Installing."
-    execOrQuit(&"nimble install {sourceName}")
-  else:
-    if verbose:
-      echo &"{packageName} installed @ {output}"
-
-task prereqs, "Install prerequisites":
-  let packages = @[
-    ("compiler", "compiler"),
-    ("anycase", "anycase"),
-    ("msgpack4nim", "msgpack4nim"),
-    ("optionsutils", "https://github.com/PMunch/nim-optionsutils")
-  ]
-  for (packageName, sourceName) in packages:
-    checkPrereq(packageName, sourceName)
 
 task gdengine_update, "update the 3.2 custom branch with changes from upstream":
 
@@ -125,7 +133,8 @@ task gdengine_update, "update the 3.2 custom branch with changes from upstream":
   for branch in gd_branches:
     execOrQuit(&"git merge {branch}")
 
-  execOrQuit(&"git push --force origin {gd_build_branch}")
+  if not ("keeplocal" in args):
+    execOrQuit(&"git push --force origin {gd_build_branch}")
 
   setCurrentDir(projDir)
 
@@ -180,17 +189,39 @@ task gd, "launches terminal with godot project\n\toptional argument for scene to
   if hostOS == "windows": discard execShellCmd &"wt -d {curDir} {gdbin} -e --path {projDir} {scn}"
   else: discard execShellCmd &"{gdbin} -e --path {projDir} {scn}"
 
+proc checkPrereq(packageName, sourceName:string, verbose:bool = true) =
+  var (output, exitCode) = execCmdEx(&"nimble path {packageName}")
+  if exitCode != 0:
+    echo &"{packageName} is not installed. Installing."
+    execOrQuit(&"nimble install {sourceName}")
+  else:
+    if verbose:
+      echo &"{packageName} installed @ {output}"
+
 task genapi, "generate the godot api bindings":
-  execOrQuit(&"nim c -r --path:{depsDir}/{depsGodotDir} {depsDir}/genapi.nim")
+  execOrQuit(&"nim c -r {gdpathFlags} {depsDir}/genapi.nim")
   var ext = if hostOS == "windows": ".exe" else : ""
   removeFile(&"{depsDir}/genapi{ext}")
+
+task prereqs, "Install prerequisites, and calls genapi task":
+  let packages = @[
+    ("compiler", "compiler"),
+    ("anycase", "anycase"),
+    ("msgpack4nim", "msgpack4nim"),
+    ("optionsutils", "https://github.com/PMunch/nim-optionsutils")
+  ]
+  for (packageName, sourceName) in packages:
+    checkPrereq(packageName, sourceName)
+
+  genapiTask()
 
 proc buildWatcher():string =
   {.cast(gcsafe).}:
     var flags = getSharedFlags()
     let dllPath = &"{dllDir}/watcher.{dllExt}"
-    if ("force" in flags) or not fileExists(&"{dllPath}") or (getLastModificationTime("watcher.nim") > getLastModificationTime(&"{dllPath}")):
-      result = execnim(&"--path:{depsDir} --path:{depsDir}/{depsGodotDir} --define:dllDir:{baseDllDir} --define:dllExt:{dllExt}", flags, &"{dllPath}", "watcher.nim")
+    let watcherPath = "gdnim/watcher.nim"
+    if ("force" in flags) or not fileExists(&"{dllPath}") or (getLastModificationTime(watcherPath) > getLastModificationTime(&"{dllPath}")):
+      result = execnim(&"{gdpathFlags} --define:dllDir:{baseDllDir} --define:dllExt:{dllExt}", flags, &"{dllPath}", watcherPath)
     else:
       result = "Watcher is unchanged"
 
@@ -225,83 +256,179 @@ task cleandll, "clean the dlls, arguments are component names, default all non-g
     echo &"rm {dllPath}"
     removeFile dllPath
 
-task gencomp, "generate a component template (nim, gdns, tscn files), pass in the component name and  base class name in snake case:":
-  var compName = args[0]
+proc getBuildSettings(): BuildSettings =
+  result.sharedFlags = getSharedFlags()
+  var settingsTable: Table[string, bool]
+  settingsTable["move"] = "move" in otherFlagsTable
+  settingsTable["newOnly"] = "force" notin result.sharedFlags
+  settingsTable["noCheck"] = "nocheck" in otherFlagsTable
+  settingsTable["noReload"] = "reload" notin result.sharedFlags
+  result.settingsTable = settingsTable
+
+proc safeDllFilePath(compName:string): string =
+  &"{dllDir}/{compName}_safe.{dllExt}"
+proc hotDllFilePath(compName:string): string =
+  &"{dllDir}/{compName}.{dllExt}"
+proc nimFilePath(compName:string): string =
+  var nimFilePath = &"{compsDir}/{compName}.nim"
+  if not fileExists(nimFilePath):
+    nimFilePath = &"{compsDir}/tools/{compName}.nim"
+  nimFilePath
+
+proc shouldBuild(compName:string, buildSettings:BuildSettings ):bool =
+  let safe = safeDllFilePath(compName)
+  let hot = hotDllFilePath(compName)
+  let nim = nimFilePath(compName)
+  result = buildSettings.settingsTable["noCheck"] or
+    (not buildSettings.settingsTable["newOnly"]) or
+    (buildSettings.settingsTable["newOnly"] and (
+      (not fileExists(hot) and not fileExists(safe)) or
+      (fileExists(safe) and getLastModificationTime(nim) > getLastModificationTime(safe)) or
+      (fileExists(hot) and not fileExists(safe) and getLastModificationTime(nim) > getLastModificationTime(hot))
+    ))
+
+proc buildComp(compName:string, buildSettings:BuildSettings):string =
+  {.cast(gcsafe).}:
+    let safe = safeDllFilePath(compName)
+    let hot = hotDllFilePath(compName)
+    let nim = nimFilePath(compName)
+
+    if not fileExists(nim):
+      result &= &"Error: '{compName}' not found in components or components/tools"
+      return
+
+    genGdns(compName)
+
+    if shouldBuild(compName, buildSettings):
+      result &= &">>> Build {compName} <<<"
+      result &= execnim(&"{gdpathFlags} --path:.", buildSettings.sharedFlags, &"{safe}", &"{nim}")
+
+    if fileExists(safe) and getLastModificationTime(nim) < getLastModificationTime(safe) and
+      (not fileExists(hot) or buildSettings.settingsTable["move"]) or buildSettings.settingsTable["noReload"]:
+      moveFile(safe, hot)
+      result &= ">>> dll moved safe to hot <<<"
+
+proc buildAllComps(res:var seq[FlowVar[string]], buildSettings:BuildSettings):int =
+  var count = 0
+  echo "building components: "
+  for compPath in walkFiles(&"{compsDir}/*.nim"):
+    inc count
+    var compName = splitFile(compPath)[1].snake
+    if shouldBuild(compName, buildSettings):
+      echo "  " & compName
+      res.add(spawn buildComp(compName, buildSettings))
+  for compPath in walkFiles(&"{compsDir}/tools/*.nim"):
+    inc count
+    var compName = splitFile(compPath)[1].snake
+    if shouldBuild(compName, buildSettings):
+      echo "  " & compName
+      res.add(spawn buildComp(compName, buildSettings))
+  count
+
+
+task gencomp, "generate a component template (nim, gdns, tscn files), pass in the component name and  base class name in snake case\n\tUsage: ./build gencomp [notscn] comp_name base_node":
+
+  var compName:string
+  var baseClassModuleName:string
+  case args.len:
+    of 2:
+      compName = args[0]
+      baseClassModuleName = args[1].tolower
+    of 3:
+      compName = args[1]
+      baseClassModuleName = args[2].tolower
+    else:
+      echo "Usage: ./build gencomp [notscn] comp_name base_node"
+      echo "Example: ./build gencomp player kinematic_body_2d"
+      echo "Example without tscn: ./build gencomp notscn effect animated_sprite"
+      quit()
+
   var compClassName = compName.pascal
-  var baseClassModuleName = args[1].tolower
   var baseClassName = baseClassModuleName.pascal
   if baseClassName.endsWith("1d") or baseClassName.endsWith("2d") or baseClassName.endsWith("3d"):
     baseClassName[^1] = 'D'
 
+  baseClassModuleName = case baseClassModuleName:
+    of "object": "objects"
+    else: baseClassModuleName
+
+  var classFilename = &"{depsDir}/godotapi/{baseClassModuleName}.nim"
+  if not fileExists(classFilename):
+    echo &"Error generating component. Could not find {classFilename}!"
+    quit()
+
   let nim = &"{compsDir}/{compName}.nim"
   if not fileExists(nim):
     var f = open(nim, fmWrite)
-    f.write(nim_template % [compName, compClassName, baseClassModuleName, baseClassName])
+    f.write(script_nim_template % [compName, compClassName, baseClassModuleName, baseClassName])
     f.close()
     echo &"generated {nim}"
+  else:
+    echo &"{nim} already exists"
+
+  if "notscn" notin args[0]:
+    let tscn = &"{tscnDir}/{compName}.tscn"
+    if not fileExists(tscn):
+      var f = open(tscn, fmWrite)
+      f.write(tscn_template % [compName, compClassName, relativePath(gdnsDir, appDir), baseClassName])
+      f.close()
+      echo &"generated {tscn}"
+    else:
+      echo &"{tscn} already exists"
 
   genGdns(compName)
 
-  let tscn = &"{tscnDir}/{compName}.tscn"
-  if not fileExists(tscn):
-    var f = open(tscn, fmWrite)
-    f.write(tscn_template % [compName, compClassName, relativePath(gdnsDir, appDir), baseClassName])
-    f.close()
-    echo &"generated {tscn}"
+  var buildSettings = getBuildSettings()
+  echo &"building {compName}"
+  echo buildComp(compName, buildSettings)
 
-proc prepBuildCompSettings(): tuple[sharedFlags:string, buildSettings:Table[string, bool]] =
-  result.sharedFlags = getSharedFlags()
-  var buildSettings: Table[string, bool]
-  buildSettings["move"] = "move" in otherFlagsTable
-  buildSettings["newOnly"] = not ("force" in result.sharedFlags)
-  buildSettings["noCheck"] = "nocheck" in otherFlagsTable
-  result.buildSettings = buildSettings
-
-proc buildComp(compName:string, sharedFlags:string, buildSettings:Table[string, bool]):string =
-  {.cast(gcsafe).}:
-    genGdns(compName)
-
-    let safeDllFilePath = &"{dllDir}/{compName}_safe.{dllExt}"
-    let hotDllFilePath = &"{dllDir}/{compName}.{dllExt}"
-    let nimFilePath = &"{compsDir}/{compName}.nim"
-
-    if not fileExists(nimFilePath):
-      result &= &"Error compiling {nimFilePath} [Not Found]"
-      return
-
-    if buildSettings["noCheck"] or
-      (not buildSettings["newOnly"]) or
-      (buildSettings["newOnly"] and (
-        (not fileExists(hotDllFilePath) and not fileExists(safeDllFilePath)) or
-        (fileExists(safeDllFilePath) and getLastModificationTime(nimFilePath) > getLastModificationTime(safeDllFilePath)) or
-        (fileExists(hotDllFilePath) and not fileExists(safeDllFilePath) and getLastModificationTime(nimFilePath) > getLastModificationTime(hotDllFilePath))
-      )):
-      result &= &">>> Build {compName} <<<"
-      result &= execnim(&"--path:{depsDir} --path:{depsDir}/{depsGodotDir} --path:.", sharedFlags, &"{safeDllFilePath}", &"{nimFilePath}")
-
-    if fileExists(safeDllFilePath) and getLastModificationTime(nimFilePath) < getLastModificationTime(safeDllFilePath) and
-      (not fileExists(hotDllFilePath) or buildSettings["move"]):
-      moveFile(safeDllFilePath, hotDllFilePath)
-      result &= ">>> dll moved safe to hot <<<"
 
 # components are named {compName}_safe.dll and
 # are loaded by the watcher.dll via resources. At runtime, the watcher.dll will copy
 # the {compName}_safe.dll to the {compName}.dll and monitor the _dlls
 # folder to see if _safe.dll is rebuilt.
 task comp, "build component and generate a gdns file\n\tno component name means all components are rebuilt\n\tmove safe to hot with 'move' or 'm' flag (e.g.) build -m target\n\t--force or --f force rebuilds\n\t--nocheck or --nc skips compile without dll check but not force rebuilt":
-  var (sharedFlags, buildSettings) = prepBuildCompSettings()
+  var buildSettings = getBuildSettings()
 
   if not (compName == ""):
     compName = compName.snake
-    echo buildComp(compName, sharedFlags, buildSettings)
+    echo buildComp(compName, buildSettings)
   else:
     # compile all the comps
     var res = newSeq[FlowVar[string]]()
-    for compPath in walkFiles(&"{compsDir}/*.nim"):
-      res.add(spawn buildComp(splitFile(compPath)[1].snake, sharedFlags, buildSettings))
+    discard buildAllComps(res, buildSettings)
     sync()
     for f in res:
       echo ^f
+
+task gentool, "generate a tool / editor plugin scaffold":
+  if args.len != 1:
+    echo "Usage: ./build gentool tool_name"
+    quit()
+
+  var compName = args[0]
+  var compClassName = compName.pascal
+
+  let nim = &"{compsDir}/tools/{compName}.nim"
+  if not fileExists(nim):
+    var f = open(nim, fmWrite)
+    f.write(tool_nim_template % [compName, compClassName])
+    f.close()
+    echo &"generated {nim}"
+  else:
+    echo &"{nim} already exists"
+
+  let cfg = &"{appDir}/addons/{compName}/plugin.cfg"
+  if not fileExists(cfg):
+    createDir(&"{appDir}/addons/{compName}")
+    var f = open(cfg, fmWrite)
+    f.write(plugin_cfg_template % [compName])
+    f.close()
+    echo &"generated {cfg}"
+  else:
+    echo &"{cfg} already exists"
+
+  genGdns(compName)
 
 task flags, "display the flags used for compiling components":
   echo ">>> Task  compiler flags <<<"
@@ -324,15 +451,12 @@ task cleanbuild, "Rebuild all":
 
   var res = newSeq[FlowVar[string]]()
 
+  echo "building watcher"
   res.add(spawn buildWatcher())
 
   var compileCount = 1
-
   # comp task
-  var (sharedFlags, buildSettings) = prepBuildCompSettings()
-  for compPath in walkFiles(&"{compsDir}/*.nim"):
-    inc compileCount
-    res.add(spawn buildComp(splitFile(compPath)[1].snake, sharedFlags, buildSettings))
+  compileCount += buildAllComps(res, getBuildSettings())
   sync()
 
   var successes = 0
@@ -353,11 +477,12 @@ task cleanbuild, "Rebuild all":
     echo "=== <<< Build Failed <<< ==="
 
 task cwatch, "Monitors the components folder for changes to recompile.":
+  echo "Monitoring components for changes.  Ctrl+C to stop"
   var lastTimes = newTable[string, Time]()
   for compPath in walkFiles(&"{compsDir}/*.nim"):
     lastTimes[compPath] = getLastModificationTime(compPath)
 
-  var (sharedFlags, buildSettings) = prepBuildCompSettings()
+  var buildSettings = getBuildSettings()
 
   while true:
     for compPath in walkFiles(&"{compsDir}/*.nim"):
@@ -366,7 +491,7 @@ task cwatch, "Monitors the components folder for changes to recompile.":
         lastTimes[compPath] = curLastTime
         var compFilename = splitFile(compPath)[1].snake
         echo &"-- Recompiling {compFilename} --"
-        echo buildComp(compFilename, sharedFlags, buildSettings)
+        echo buildComp(compFilename, buildSettings)
     sleep cwatch_interval
 
 task diagnostic, "Displays code that contributes to your dll size. Pass in the comp name as an argument: ./build diagnostic comp_name":
@@ -380,3 +505,6 @@ task help, "display list of tasks":
   echo "Call build with a task:"
   for i in 0..<tasks.len:
     echo "  ", tasks[i].task_name, " : ", tasks[i].description
+
+  echo "\nAdditional flags"
+  echo "  --ini:build.ini : sets the config file"
